requirements.txt
contourpy==1.3.3
cycler==0.12.1
duckdb==1.4.1
fonttools==4.60.1
kiwisolver==1.4.9
matplotlib==3.10.7
narwhals==2.11.0
numpy==2.3.3
packaging==25.0
pandas==2.3.3
pillow==12.0.0
plotly==6.4.0
pyparsing==3.2.5
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
scipy==1.16.2
six==1.17.0
tqdm==4.67.1
tzdata==2025.2
